migrate variant header
