+++ common.scss
@@ -1,11 +1,7 @@
-/*
-
-You can set up a global style here that is commonly used in each stanza.
-
-Example:
-
-h1 {
-  font-size: 24px;
+main {
+  font-family: 'Roboto Condensed', sans-serif;
 }
 
-*/
+.table {
+  border-collapse: collapse;
+}
