+++ lib/uniq.js
@@ -1,0 +1,1 @@
+export default array => array.filter((elem, index, self) => self.indexOf(elem) === index)
