+++ stanzas/variant-header/README.md
@@ -1,0 +1,3 @@
+# Variant header
+
+Stanza description goes here. Edit `stanzas/variant-header/README.md` to update.
