+++ stanzas/variant-header/style.scss
@@ -1,0 +1,22 @@
+@use 'common';
+
+main {
+  padding: 1rem 2rem;
+}
+
+.table {
+  &.variant-header {
+    thead tr th {
+      font-size: 12px;
+      text-align: var(--text-align);
+      color: #1f5065;
+    }
+    tbody td {
+      font-size: 14px;
+      text-align: var(--text-align);
+      a {
+        color: #ffffff;
+      }
+    }
+  }
+}
