+++ stanzas/variant-header/templates/fetch_xrefs.rq
