--- variant_header/assets/style.css
@@ -1,21 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Roboto+Condensed:300,400,700,900");
-
-.table {
-  border-collapse: collapse;
-  font-family: 'Roboto Condensed', sans-serif;
-}
-
-.table.header-xref > thead > tr > th {
-  font-size: 12px;
-  color: #1f5065;
-  text-align: center;
-}
-
-.table.header-xref > tbody > tr > td {
-  font-size: 14px;
-  text-align: center;
-}
-
-.table.header-xref > tbody > tr > td > a {
-  color: #ffffff;
-}
