--- variant_header/metadata.json
@@ -1,34 +1,0 @@
-{
-  "@context": {
-    "stanza": "http://togostanza.org/resource/stanza#"
-  },
-  "@id": "variant_header",
-  "stanza:label": "Variant / Header",
-  "stanza:definition": "Display header information",
-  "stanza:parameter": [
-    {
-      "stanza:key": "tgv_id",
-      "stanza:example": "tgv219804",
-      "stanza:description": "TogoVar ID",
-      "stanza:required": true
-    },
-    {
-      "stanza:key": "ep",
-      "stanza:example": "https://togovar.biosciencedbc.jp/sparql",
-      "stanza:description": "SPARQL endpoint URL",
-      "stanza:required": false
-    }
-  ],
-  "stanza:usage": "<togostanza-variant_header tgv_id=\"tgv219804\"></togostanza-variant_header>",
-  "stanza:type": "Stanza",
-  "stanza:context": "",
-  "stanza:display": "",
-  "stanza:provider": "provider of this stanza",
-  "stanza:license": "",
-  "stanza:author": "Daisuke Satoh",
-  "stanza:address": "daisuke.satoh@lifematics.co.jp",
-  "stanza:contributor": [
-  ],
-  "stanza:created": "2019-06-07",
-  "stanza:updated": "2020-07-17"
-}
